feat: reorganize tools into confluence and jira directories; add CQL and page tools
