+++ src/tools/confluence/registerPageTools.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ConfluenceClient } from "../clients/confluenceClient";
+import { ConfluenceClient } from "../../clients/confluenceClient";
 
 export function registerPageTools(server: any, config: any) {
   server.tool(
